Simplify and spread logic into separate file
