+++ lib/core/src/server/config/webpack.config.default.js
@@ -1,44 +1,13 @@
-import autoprefixer from 'autoprefixer';
-import findUp from 'find-up';
+import getStyleConfig from './style.config';
 
 export function createDefaultWebpackConfig(storybookBaseConfig) {
-  const postcssConfig = findUp.sync('postcss.config.js', {
-    cwd: storybookBaseConfig.context,
-  });
-
   return {
     ...storybookBaseConfig,
     module: {
       ...storybookBaseConfig.module,
       rules: [
         ...storybookBaseConfig.module.rules,
-        {
-          test: /\.css$/,
-          use: [
-            require.resolve('style-loader'),
-            {
-              loader: require.resolve('css-loader'),
-              options: {
-                importLoaders: 1,
-              },
-            },
-            {
-              loader: require.resolve('postcss-loader'),
-              options: postcssConfig
-                ? {
-                    postcssConfig,
-                  }
-                : {
-                    plugins: () => [
-                      require('postcss-flexbugs-fixes'), // eslint-disable-line global-require
-                      autoprefixer({
-                        flexbox: 'no-2009',
-                      }),
-                    ],
-                  },
-            },
-          ],
-        },
+        getStyleConfig(storybookBaseConfig),
         {
           test: /\.(svg|ico|jpg|jpeg|png|gif|eot|otf|webp|ttf|woff|woff2)(\?.*)?$/,
           loader: require.resolve('file-loader'),
